+++ crossfitApi/app/Http/Controllers/ExerciseController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Exercise;
 use Illuminate\Http\Request;
 use App\Libs\ResultResponse;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
 
 class ExerciseController extends Controller
 {
@@ -27,17 +29,24 @@ class ExerciseController extends Controller
     public function store(Request $request)
     {
         $resultResponse = new ResultResponse();
-        try {
-            $newPlan = new Exercise([
-                'ej_nombre' => $request->get('ej_nombre'),
-            ]);
-            $newPlan->save();
-            $resultResponse->setData($newPlan);
-            $resultResponse->setStatusCode(ResultResponse::SUCCESS_CODE);
-            $resultResponse->setMessage(ResultResponse::TXT_SUCCESS_CODE);
-        } catch (\Exception $e) {
+        $validation = $this->validateRequest($request);
+        if ($validation->fails()) {
+            $errors = $validation->errors()->all();
             $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
-            $resultResponse->setMessage(ResultResponse::TXT_ERROR_CODE);
+            $resultResponse->setMessage(implode(', ', $errors));
+        } else {
+            try {
+                $newPlan = new Exercise([
+                    'ej_nombre' => $request->get('ej_nombre'),
+                ]);
+                $newPlan->save();
+                $resultResponse->setData($newPlan);
+                $resultResponse->setStatusCode(ResultResponse::SUCCESS_CODE);
+                $resultResponse->setMessage(ResultResponse::TXT_SUCCESS_CODE);
+            } catch (\Exception $e) {
+                $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
+                $resultResponse->setMessage(ResultResponse::TXT_ERROR_CODE);
+            }
         }
         return response()->json($resultResponse);
     }
@@ -48,14 +57,21 @@ class ExerciseController extends Controller
     public function show($id)
     {
         $resultResponse = new ResultResponse();
-        try {
-            $exercise = Exercise::findOrFail($id);
-            $resultResponse->setData($exercise);
-            $resultResponse->setStatusCode(ResultResponse::SUCCESS_CODE);
-            $resultResponse->setMessage(ResultResponse::TXT_SUCCESS_CODE);
-        } catch (\Exception $e) {
+        $validation = $this->validateId($id);
+        if ($validation->fails()) {
+            $errors = $validation->errors()->all();
             $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
-            $resultResponse->setMessage(ResultResponse::TXT_ERROR_CODE);
+            $resultResponse->setMessage(implode(', ', $errors));
+        } else {
+            try {
+                $exercise = Exercise::findOrFail($id);
+                $resultResponse->setData($exercise);
+                $resultResponse->setStatusCode(ResultResponse::SUCCESS_CODE);
+                $resultResponse->setMessage(ResultResponse::TXT_SUCCESS_CODE);
+            } catch (\Exception $e) {
+                $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
+                $resultResponse->setMessage(ResultResponse::TXT_ERROR_CODE);
+            }
         }
         return response()->json($resultResponse);
     }
@@ -66,17 +82,23 @@ class ExerciseController extends Controller
     public function update(Request $request, $id)
     {
         $resultResponse = new ResultResponse();
-
-        try {
-            $exercise = Exercise::findOrFail($id);
-            $exercise->ej_nombre = $request->get('ej_nombre', $exercise->ej_nombre);
-            $exercise->save();
-            $resultResponse->setData($exercise);
-            $resultResponse->setStatusCode(ResultResponse::SUCCESS_CODE);
-            $resultResponse->setMessage(ResultResponse::TXT_SUCCESS_CODE);
-        } catch (\Exception $e) {
+        $validation = $this->validateRequest($request);
+        if ($validation->fails()) {
+            $errors = $validation->errors()->all();
             $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
-            $resultResponse->setMessage(ResultResponse::TXT_ERROR_CODE);
+            $resultResponse->setMessage(implode(', ', $errors));
+        } else {
+            try {
+                $exercise = Exercise::findOrFail($id);
+                $exercise->ej_nombre = $request->get('ej_nombre', $exercise->ej_nombre);
+                $exercise->save();
+                $resultResponse->setData($exercise);
+                $resultResponse->setStatusCode(ResultResponse::SUCCESS_CODE);
+                $resultResponse->setMessage(ResultResponse::TXT_SUCCESS_CODE);
+            } catch (\Exception $e) {
+                $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
+                $resultResponse->setMessage(ResultResponse::TXT_ERROR_CODE);
+            }
         }
         return response()->json($resultResponse);
     }
@@ -107,16 +129,75 @@ class ExerciseController extends Controller
     public function destroy($id)
     {
         $resultResponse = new ResultResponse();
+        $validation = $this->validateId($id);
+        if ($validation->fails()) {
+            $errors = $validation->errors()->all();
+            $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
+            $resultResponse->setMessage(implode(', ', $errors));
+        } else {
+            try {
+                $exercise = Exercise::findOrFail($id);
+                $exercise->delete();
+                $resultResponse->setData($exercise);
+                $resultResponse->setStatusCode(ResultResponse::SUCCESS_CODE);
+                $resultResponse->setMessage(ResultResponse::TXT_SUCCESS_CODE);
+            } catch (\Exception $e) {
+                $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
+                $resultResponse->setMessage(ResultResponse::TXT_ERROR_CODE);
+            }
+        }
+        return response()->json($resultResponse);
+    }
+    public function search(Request $request)
+    {
+        $resultResponse = new ResultResponse();
         try {
-            $exercise = Exercise::findOrFail($id);
-            $exercise->delete();
-            $resultResponse->setData($exercise);
+            // Obtener todos los ejercicios
+            $exercises = Exercise::all();
+
+            // Aplicar búsqueda si se proporciona un término de búsqueda
+            if ($request->has('search_term')) {
+                $searchTerm = strtolower($request->input('search_term'));
+
+                // Filtrar ejercicios que coincidan con el término de búsqueda en cualquiera de los campos
+                $exercises = $exercises->filter(function ($exercise) use ($searchTerm) {
+                    foreach ($exercise->toArray() as $field => $value) {
+                        if (is_string($value) && str_contains(strtolower($value), $searchTerm)) {
+                            return true;
+                        }
+                    }
+                    return false;
+                });
+            }
+
+            $resultResponse->setData($exercises);
             $resultResponse->setStatusCode(ResultResponse::SUCCESS_CODE);
             $resultResponse->setMessage(ResultResponse::TXT_SUCCESS_CODE);
         } catch (\Exception $e) {
             $resultResponse->setStatusCode(ResultResponse::ERROR_CODE);
             $resultResponse->setMessage(ResultResponse::TXT_ERROR_CODE);
         }
+
         return response()->json($resultResponse);
+    }
+
+    private function validateRequest(Request $request)
+    {
+        // Reglas de validación
+        $rules = [
+            'ej_nombre' => 'required|string|max:255',
+        ];
+        return Validator::make($request->all(), $rules);
+    }
+
+    private function validateId($id)
+    {
+        return Validator::make(['id' => $id], [
+            'id' => [
+                'required',
+                'integer',
+                Rule::exists('exercises', 'id'),
+            ],
+        ]);
     }
 }
